fixes and adds tests
