+++ scripts/build.js
@@ -4,6 +4,7 @@ import * as esbuild from 'esbuild';
 import { copyFile, mkdir } from 'fs/promises';
 import { dirname } from 'path';
 import { fileURLToPath } from 'url';
+import {green} from "barva";
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const isWatch = process.argv.includes('--watch');
@@ -12,8 +13,9 @@ async function copyAssets() {
   try {
     await mkdir('dist/assets', { recursive: true });
     await copyFile('src/assets/404.html', 'dist/assets/404.html');
-    await copyFile('src/assets/listing.html', 'dist/assets/listing.html');
     await copyFile('src/assets/favicon.ico', 'dist/assets/favicon.ico');
+    await copyFile('src/assets/img.png', 'dist/assets/img.png');
+    await copyFile('src/assets/listing.html', 'dist/assets/listing.html');
   } catch (err) {
     console.error('Error copying assets:', err);
   }
@@ -33,7 +35,6 @@ const buildOptions = {
   },
   external: [
     'child_process',
-    'figlet',
     'fs',
     'http',
     'node',
@@ -53,5 +54,5 @@ if (isWatch) {
   // Build mode
   await esbuild.build(buildOptions);
   await copyAssets();
-  console.log('Build completed');
+  console.log(green`\n✓ Build completed\n`);
 }
