+++ test/spec/cli.spec.js
@@ -6,7 +6,7 @@ import { jest } from '@jest/globals';
 import { magenta, green, grey, red } from 'barva';
 
 // Mock dependencies
-jest.mock('../../src/server/index.js', () => ({
+jest.mock('../../src/server/create-server.js', () => ({
   createServer: jest.fn(() => ({
     listen: jest.fn(),
     close: jest.fn(cb => cb && cb())
@@ -53,22 +53,22 @@ describe('CLI Module', () => {
     it('should display the ASCII logo with correct styling', async () => {
       // We need to use dynamic import to allow mocking before import
       const config = { version: '1.0.0', address: 'localhost', port: 8080 };
-      
+
       // Get the displaySplashScreen function through module internals
       const displaySplashScreen = jest.fn();
       jest.doMock('../../src/cli/index.js', () => ({
         __esModule: true,
         displaySplashScreen
       }));
-      
+
       // Now we can verify the logo is displayed correctly by checking console.log calls
-      const mockConfig = { 
-        version: '1.0.0', 
-        address: 'localhost', 
+      const mockConfig = {
+        version: '1.0.0',
+        address: 'localhost',
         port: 8080,
         portChanged: false
       };
-      
+
       // Manual testing with a simplified function that mimics displaySplashScreen
       const testDisplayLogo = () => {
         const logo = `
@@ -83,9 +83,9 @@ describe('CLI Module', () => {
 `;
         console.log(magenta`${logo}`);
       };
-      
+
       testDisplayLogo();
-      
+
       // Verify console.log was called with the magenta-colored logo
       expect(mockConsoleLog).toHaveBeenCalled();
       const calls = mockConsoleLog.mock.calls;
